Update: emit DomainModel instead of string on 'create site button' click
