+++ WordPress/src/main/java/org/wordpress/android/ui/sitecreation/domains/DomainsScreenListener.kt
@@ -1,5 +1,12 @@
 package org.wordpress.android.ui.sitecreation.domains
 
 interface DomainsScreenListener {
-    fun onDomainSelected(domain: String)
+    fun onDomainSelected(domain: DomainModel)
 }
+
+data class DomainModel(
+    val domainName: String,
+    val isFree: Boolean,
+    val cost: String,
+    val productId: Int,
+)
